Word2VecSynonyms - Moved parameters maxResult and accuracy + tests
Since the introduction of the Word2VecSynonymProviderFactory, we keep in memory a single
instance of the Word2VecSynonymProvider. On the other hand, we may have multiple instances of the
Word2VecSynonymFilterFactory with different parameters maxResult and accuracy.
We cannot embed the parameters maxResult and accuracy in the SynonymProvider but we need to
pass it every time we need to perform a query